Add new hooks, that takes functions from app.tsx. More clean
