+++ src/hooks/useDarkMode.ts
@@ -1,0 +1,33 @@
+import { useState, useEffect } from 'react';
+
+export const useDarkMode = () => {
+  const [isDarkMode, setIsDarkMode] = useState(false);
+
+  // Cargar tema del localStorage al montar
+  useEffect(() => {
+    const savedTheme = localStorage.getItem('theme');
+    if (savedTheme === 'dark') {
+      setIsDarkMode(true);
+      document.documentElement.setAttribute('data-theme', 'dark');
+    }
+  }, []);
+
+  // Manejar toggle de tema
+  const handleThemeToggle = () => {
+    const newTheme = !isDarkMode;
+    setIsDarkMode(newTheme);
+
+    if (newTheme) {
+      document.documentElement.setAttribute('data-theme', 'dark');
+      localStorage.setItem('theme', 'dark');
+    } else {
+      document.documentElement.removeAttribute('data-theme');
+      localStorage.setItem('theme', 'light');
+    }
+  };
+
+  return {
+    isDarkMode,
+    handleThemeToggle,
+  };
+};
