+++ tsconfig.app.json
@@ -24,6 +24,7 @@
     //"erasableSyntaxOnly": true,
     "noFallthroughCasesInSwitch": true,
     //"noUncheckedSideEffectImports": true
+    "forceConsistentCasingInFileNames": true
   },
   "include": ["src"]
 }
